fixed quickselect bug, made it run faster, improved tools

+++ tools/compare.sh
@@ -40,41 +40,44 @@ testTSP()
 
 if [ $# != 2 -a $# != 3 ]
 then
-	echo 'Usage: $0 number_of_test_samples number_of_test_repetitions [best]'
+	echo 'Usage: $0 number_of_test_samples number_of_test_repetitions [old]'
 	exit 1
 fi
 
-echo "Generating $2 test inputs of size $1..."
-for i in $(seq 1 $2)
-do
-	$TESTGEN_BIN $1 > testIn$i.txt
-done
-
 if [ $# != 3 ]
 then
-	echo
-	echo 'Testing nearest neighbor:'
-	testTSP $2 n
-
-	echo
-	echo 'Testing greedy:'
-	testTSP $2 g
-
-	echo
-	echo 'Testing randomized greedy:'
-	testTSP $2 G
-
-	echo
-	echo 'Testing greedy + 2opt:'
-	testTSP $2 1
-
-	echo
-	echo 'Testing randomized greedy + 2opt:'
-	testTSP $2 2
+	echo "Generating $2 test inputs of size $1..."
+	for i in $(seq 1 $2)
+	do
+		$TESTGEN_BIN $1 > testIn$i.txt
+	done
 fi
+
+echo
+echo 'Testing nearest neighbor:'
+testTSP $2 n
+
+echo
+echo 'Testing greedy:'
+testTSP $2 g
+
+echo
+echo 'Testing randomized greedy:'
+testTSP $2 G
+
+echo
+echo 'Testing greedy + 2opt:'
+testTSP $2 1
+
+echo
+echo 'Testing randomized greedy + 2opt:'
+testTSP $2 2
 
 echo
 echo 'Testing iterative randomized greedy + 2opt:'
 testTSP $2 3
 
-rm testIn*.txt
+if [ $# != 3 ]
+then
+	rm testIn*.txt
+fi
